+++ src/main/java/io/apiman/cli/core/api/Version11xServerApi.java
@@ -25,11 +25,11 @@ import retrofit.http.*;
 import java.util.List;
 
 /**
- * Legacy support for apiman 1.1.9.
+ * Legacy support for apiman 1.1.x.
  *
  * @author Pete Cornish {@literal <outofcoffee@gmail.com>}
  */
-public interface ServiceApi {
+public interface Version11xServerApi {
     @POST("/organizations/{orgName}/services")
     Response create(@Path("orgName") String orgName, @Body Api api);
 
@@ -50,4 +50,8 @@ public interface ServiceApi {
     @GET("/organizations/{orgName}/services/{serviceName}/versions/{version}/policies")
     List<ApiPolicy> fetchPolicies(@Path("orgName") String orgName, @Path("serviceName") String serviceName,
                                   @Path("version") String version);
+
+    @PUT("/organizations/{orgName}/services/{serviceName}/versions/{version}/policies/{policyId}")
+    Response configurePolicy(@Path("orgName") String orgName, @Path("serviceName") String serviceName,
+                             @Path("version") String version, @Path("policyId") Long policyId, @Body ApiPolicy policyConfig);
 }