+++ src/main/java/io/apiman/cli/server/binding/ServerApiBindings.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *    http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,22 +14,21 @@
  * limitations under the License.
  */
 
-package io.apiman.cli.core.api;
+package io.apiman.cli.server.binding;
 
-import io.apiman.cli.core.common.action.ModelAction;
-import io.apiman.cli.core.api.model.Api;
+import io.apiman.cli.core.common.model.ServerVersion;
 
 /**
+ * Convenience methods for instantiating {@link ServerApiBinding}s.
+ *
  * @author Pete Cornish {@literal <outofcoffee@gmail.com>}
  */
-public interface ApiMixin extends ModelAction<Api, Version12xServerApi> {
-    @Override
-    default Class<Version12xServerApi> getApiClass() {
-        return Version12xServerApi.class;
+public class ServerApiBindings {
+    public static ServerApiBinding boundTo(Class<?> apiClass) {
+        return new ServerApiBindingImpl(apiClass);
     }
 
-    @Override
-    default Class<Api> getModelClass() {
-        return Api.class;
+    public static ServerApiBinding boundTo(Class<?> apiClass, ServerVersion serverVersion) {
+        return new ServerApiBindingImpl(apiClass, serverVersion);
     }
 }