+++ src/test/java/io/apiman/cli/support/TestModel.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *    http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,45 +14,34 @@
  * limitations under the License.
  */
 
-package io.apiman.cli.core.api.model;
+package io.apiman.cli.support;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-
-import java.util.List;
+import com.google.common.base.Objects;
 
 /**
- * Legacy support for apiman 1.1.x.
+ * Model for tests.
  *
  * @author Pete Cornish {@literal <outofcoffee@gmail.com>}
  */
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class ServiceConfig {
+public class TestModel {
     @JsonProperty
-    private String endpoint;
+    private String key;
 
-    @JsonProperty
-    private String endpointType;
-
-    @JsonProperty
-    private boolean publicService;
-
-    @JsonProperty
-    private List<ApiGateway> gateways;
-
-    public ServiceConfig() {
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        TestModel testModel = (TestModel) o;
+        return Objects.equal(key, testModel.key);
     }
 
-    public ServiceConfig(String endpoint, String endpointType, boolean publicService, List<ApiGateway> gateways) {
-        this.endpoint = endpoint;
-        this.endpointType = endpointType;
-        this.publicService = publicService;
-        this.gateways = gateways;
-    }
-
-    public void setPublicService(boolean publicService) {
-        this.publicService = publicService;
+    @Override
+    public int hashCode() {
+        return Objects.hashCode(key);
     }
 }